+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes in version 0.3.3
 * Added icon for *sharedcontent* plugin of *django-fluent-contents*.
 * Fixed ``KeyError`` when a model has add support, but no edit support.
 * Fixed icon layout when a model has no permissions to add/edit.
+* Fixed welcome text in personal module, remove pages link if the user has no permission to edit pages.
 
 
 Changes in version 0.3.2